Move code involved with album art generation into its own file.
